# backend/controllers/userController.js
const jwt = require('jsonwebtoken')
const brcypt = require('bcryptjs')
const asyncHandler = require('express-async-handler')
const User = require('../models/userModel')
const bcrypt = require('bcryptjs/dist/bcrypt')

//  @desc    Register user
//  @route   POST /api/users
//  @access  Public
const registerUser = asyncHandler(async (req, res) => {
    const { name, email, password } = req.body
    
    // Form validation
    if(!name || !email || !password) {
        res.status(400)
        throw new Error("Please add all fields")
    }

    // Check if user exists
    const userExists = await User.findOne({email})

    if (userExists) {
        res.status(400)
        throw new Error('User already exists')
    }

    // Hash password
    const salt = await bcrypt.genSalt(10)
    const hashedPassword = await brcypt.hash(password, salt)

    // Create user
    const user = await User.create({
        name,
        email,
        password: hashedPassword
    })

    if(user) {
        res.status(201).json({
            _id: user.id,
            name: user.name,
            email: user.email,
            token: generateToken(user._id)
        })
    } else {
        res.status(400)
        throw new Error('Invalid user data')
    }

})

//  @desc    Authenticate a user
//  @route   POST /api/users/login
//  @access  Public
const loginUser = asyncHandler(async (req, res) => {
    
    // Form validation
    const { email, password } = req.body
    if(!email || !password) {
        res.status(400)
        throw new Error("Please add all fields")
    }

    // Check for user email
    const user = await User.findOne({email})

    // Password check
    if(user && (await brcypt.compare(password, user.password))) {
        res.status(200).json({
            _id: user.id,
            name: user.name,
            email: user.email,
            token: generateToken(user._id)
        })
    } else {
        res.status(400)
        throw new Error('Invalid credentials')  
    }
})

//  @desc    Get user data
//  @route   GET /api/users/me
//  @access  Private
const getMe = asyncHandler(async (req, res) => {
    res.status(200).json(req.user)
})

//  Generate JWT
const generateToken = (id) => {
    return jwt.sign({ id }, process.env.JWT_SECRET, {
        expiresIn: '24h',
    })
}

module.exports = {
    registerUser,
    loginUser,
    getMe,
}